LayoutTests/inspector/debugger/command-line-api-exception-nested-catch.html

<!doctype html>
<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script type="text/javascript" src="./resources/exceptions.js"></script>
<script>
function triggerException()
{
    // We expect uncaught exceptions, so avoid logs for them.
    window.onerror = function(){};
    setTimeout(nestedCatchBlocks, 0);
}

function test()
{
    WebInspector.debuggerManager.allExceptionsBreakpoint.disabled = false;

    function dumpCommandLineAPIValue(prefix) {
        WebInspector.runtimeManager.evaluateInInspectedWindow("$exception", {objectGroup: "test", includeCommandLineAPI: true, doNotPauseOnExceptionsAndMuteConsole: true}, function(result, wasThrown) {
            InspectorTest.log(prefix + ": $exception => " + result.description);
        });
    }

    function checkIfExceptionValueMatchesVariable(varName) {
        WebInspector.runtimeManager.evaluateInInspectedWindow("$exception === " + varName, {objectGroup: "test", includeCommandLineAPI: true, doNotPauseOnExceptionsAndMuteConsole: true}, function(result, wasThrown) {
            InspectorTest.log("  CATCH: $exception === " + varName + " ? " + result.description);
        });
    }

    var done = false;
    var phase = 0;

    WebInspector.debuggerManager.addEventListener(WebInspector.DebuggerManager.Event.CallFramesDidChange, function(event) {
        if (done)
            return;

        phase++;

        // Skip past the first pause to the second pause.
        if (phase === 1) {
            dumpCommandLineAPIValue("OUTER 1");
            WebInspector.debuggerManager.resume();
            return;
        }

        // Paused on the exception in the inner try, step over to get into the inner catch.
        if (phase === 2) {
            dumpCommandLineAPIValue("INNER 1");
            WebInspector.debuggerManager.stepOver();
            return;
        }

        // Paused in the inner catch, verify $exception is "inner exception".
        if (phase === 3) {
            dumpCommandLineAPIValue("INNER 2");
            checkIfExceptionValueMatchesVariable("e2");
            WebInspector.debuggerManager.stepOver();
            return;
        }
        
        // Stepped into the outer catch, verify $exception is "outer exception".
        if (phase === 4) {
            dumpCommandLineAPIValue("OUTER 2");
            checkIfExceptionValueMatchesVariable("e1");
            WebInspector.debuggerManager.resume();
            return;
        }
    });

    WebInspector.debuggerManager.addEventListener(WebInspector.DebuggerManager.Event.Resumed, function(event) {
        done = true;
        dumpCommandLineAPIValue("AFTER  ");
        InspectorTest.completeTest();
    });

    dumpCommandLineAPIValue("BEFORE ");
    InspectorTest.evaluateInPage("triggerException()");
}
</script>
</head>
<body onload="runTest()">
    <p>Checks that <code>$exception</code> is the value of the current exception, even in nested catch blocks.</p>
</body>
</html>
